modificaciones para responsive en todo el CSS. Se modificaron y pasaron las funciones al data.js. Se añadió la función de filtrado en Locations
