+++ src/characters.html
@@ -62,7 +62,7 @@
       </tbody>
     </table>
     
-    <script src="characters.js" type="module"></script>
+    
     <script src="main.js" type="module"></script>
 </body>
 </html>
