+++ src/fundadores.html
@@ -58,7 +58,7 @@
                 <button class="close-modal">¡Wow, gracias!</button>
             </div>
         </div>
-      <script src="fundadores.js" type="module"></script>
+      <script src="main.js" type="module"></script>
   </body>
 </ht>
