+++ src/novato.css
@@ -6,21 +6,13 @@
     margin-top: 10%;
     margin-left: 10%;
     }
-@media only screen and (max-width: 768px) {/*y esto*/
-        /* For mobile phones: */
-        .frame {
-        display: flex;
-        margin-top: 5%;
-        flex-flow: column wrap;
-        }
-    }
 .History {
     width: 70%;
     background-image: url("Imágenes/newbiepage/newbie01.png");
     background-size: contain;
     background-position: right;
     overflow: hidden;
-    }  
+    }
 .Movies {
     width: 70%;
     background-image: url("Imágenes/newbiepage/newbie02.png");
@@ -52,7 +44,6 @@
     background-size: contain;
     overflow: hidden;
 }
-
 .secondBtn {
     width: 50%;
     cursor: pointer;
@@ -63,14 +54,13 @@
     margin: center;
     margin-left: 25%;
     margin-right: 10%;
-    margin-top: 7%;
+    margin-top: 2%;
     color: rgb(67, 21, 91);
 }
 .secondBtn:hover {
-    background: #ee9547;
+    background: #EE9547;
     color: #fff;
   }
-
 .modal {
     /* modals are hidden by default */
     display: none;
@@ -90,7 +80,7 @@
 }
 .modal-body {
     padding: 20px;
-    background: #b7e1f4;
+    background: #B7E1F4;
     font-family: "Inconsolata", sans-serif;
 }
 .modal-open {
@@ -99,8 +89,39 @@
 }
 .close-modal {
     margin-top: -25%;
-}    
-    
+}
+@media screen and (max-width: 768px) {
+    .navbar {
+        top: 0px;
+        width: 100%;
+        height: 100px;
+        border: 3px blueviolet;
+        margin-right: 1%;
+        background-image: url("Imágenes/paisaje.png");
+      }
+    .block {
+      display: flex;
+      flex-direction: column;
+      margin-left: -5%;
+    }
+    .frame {
+      width: 80%;
+      margin-top: 10%;
+    }
+    .modal{
+        width: 60%;
+        font-size: 40%;
+    }
+    .close-modal{
+        margin-top: 5%;
+    }
+  }
+
+
+
+
+
+
