+++ src/novato.html
@@ -40,6 +40,5 @@
         <button class="close-modal">Wow thanks!</button>
       </div>
     <script src="main.js" type="module"></script>
-    <script src="novato.js" type="module"></script>
   </body>
 </html>
