+++ src/peliculas.html
@@ -46,6 +46,5 @@
       </table>
 
   <script src="main.js" type="module"></script>
-  <script src="data.js" type="module"></script>
 </body>
 </html>
