+++ src/style.css
@@ -1,4 +1,5 @@
 @import url("https://fonts.googleapis.com/css?family=Jockey+One%7CInconsolata");
+
 .navbar {
   top: 0px;
   width: 1500px;
@@ -8,17 +9,15 @@
   background-image: url("Imágenes/paisaje.png");
 }
 html {
-  /*qué es esto*/
   height: 100%;
 }
 main {
-  /*y esto*/
   height: 100%;
 }
 body {
   width: 100%;
   height: 100%;
-  background: #093f68;
+  background: #0e0b5aef;
 }
 header {
   width: 100%;
@@ -42,7 +41,7 @@ ul li a {
   text-decoration: none;
 }
 ul li:hover {
-  background: #f1cd73;
+  background: #F1CD73;
 }
 .container {
   display: flex;
@@ -66,27 +65,11 @@ ul li:hover {
   border-radius: 8%;
   background-image: url("Imágenes/fandom.png");
 }
-/*button {
-  width: 20%;
-  padding: 14px;
-  margin-top: 48%;
-  margin-left: 40%;
-  /*margin-right: 48%;
-        margin-bottom: 25%;
-  background-color: #3cd1ef;
-  border: none;
-  cursor: pointer;
-  border-radius: 10px;
-  font-family: "Jockey One", sans-serif;
-  font-size: 20px;
-}*/
 #newbieBtn {
-  width: 20%;
+  width: 30%;
   padding: 14px;
-  margin-top: 48%;
+  margin-top: 55%;
   margin-left: 40%;
-  /*margin-right: 48%;
-        margin-bottom: 25%;*/
   background-color: #3CD1EF;
   border: none;
   cursor: pointer;
@@ -95,17 +78,14 @@ ul li:hover {
   font-size: 20px;
 }
 #newbieBtn:hover {
-  background: #ee9547;
+  background: #EE9547;
   color: #fff;
 }
-
 #fandomBtn {
-  width: 20%;
+  width: 30%;
   padding: 14px;
-  margin-top: 48%;
+  margin-top: 55%;
   margin-left: 40%;
-  /*margin-right: 48%;
-        margin-bottom: 25%;*/
   background-color: #3CD1EF;
   border: none;
   cursor: pointer;
@@ -114,25 +94,48 @@ ul li:hover {
   font-size: 20px;
 }
 #fandomBtn:hover {
-  background: #ee9547;
+  background: #EE9547;
   color: #fff;
 }
-
 @media only screen and (max-width: 768px) {
-  /*y esto*/
-  /* For mobile phones: */
+  .navbar {
+    top: 0px;
+    width: 100%;
+    height: 100px;
+    border: 3px blueviolet;
+    margin-right: 1%;
+    background-image: url("Imágenes/paisaje.png");
+  }
+    .body{
+      overflow: hidden;
+    }
   .block {
     display: flex;
     width: 100%;
     margin-top: 2%;
+    height: 40%;
   }
   .container {
     display: flex;
     margin-top: 5%;
     flex-flow: column wrap;
+    height: 100%;
+  }
+  #newbieBtn {
+    margin-top: 42%;
+  }
+  #fandomBtn {
+    margin-top: 42%;
   }
   .button {
     display: flex;
     width: 70%;
+    margin-top: -50%;
+  }
+  .newbie{
+    background-size: 100%;
+  }
+  .fandom{
+    background-size: 100%;
   }
 }
