+++ src/vehiculos.css
@@ -1,9 +1,11 @@
 .block{
     width: 50%;
     height: 100%;
+    margin-top: -5%;
 }
 h1{
     text-align: center;
+  font-family: "Jockey One", sans-serif;
 }
 table {
     width: 80%;
@@ -30,7 +32,6 @@ table {
   }
   #locationsTableColumnTwo {
     width: 50%;
-    margin-top: 15%;
     margin-left: 10%;
     text-align: center;
   }
@@ -47,7 +48,6 @@ table {
   }
   #vehiclesTableColumnTwo {
     width: 50%;
-    margin-top: 15%;
     margin-left: 10%;
     text-align: center;
     width: 250px;
@@ -64,4 +64,32 @@ table {
   }
   td{
     vertical-align: top;
+  }
+  label {
+    font-family: "Jockey One", sans-serif;
+    font-size: 120%;
+    color: #fff;
+  }
+  .vehicleName {
+    font-family: "Jockey One", sans-serif;
+    font-size: 120%;
+  }
+  #climateSelect {
+    font-family: "Jockey One", sans-serif;
+    font-size: 120%;
+    background-color: #3CD1EF;
+  }
+  #refresh {
+    flex-direction: row;
+    flex-wrap: wrap;
+    margin: auto;
+    margin-left: 0.2%;
+    display: inline-flex;
+    background-color: #3CD1EF;
+    font-family: "Jockey One", sans-serif;
+    font-size: 120%;
+  }
+  #refresh:hover {
+  background-color: #ee9547;
+  color: #fff;
   }
